+++ apps/builder/src/builder/panels/properties/editors/index.ts
@@ -1,5 +1,3 @@
-export { SelectEditor } from "./SelectEditor";
-export { ComboBoxEditor } from "./ComboBoxEditor";
 export { ComboBoxItemEditor } from "./ComboBoxItemEditor";
 export { CheckboxEditor } from "./CheckboxEditor";
 export { RadioEditor } from "./RadioEditor";
@@ -23,9 +21,7 @@ export { ColumnGroupEditor } from "./ColumnGroupEditor";
 export { RowEditor } from "./RowEditor";
 export { CellEditor } from "./CellEditor";
 export { CardEditor } from "./CardEditor";
-export { TagGroupEditor } from "./TagGroupEditor";
 export { TagEditor } from "./TagEditor";
-export { TreeEditor } from "./TreeEditor";
 export { TreeItemEditor } from "./TreeItemEditor";
 export { FieldEditor } from "./FieldEditor";
 // TODO: Implement these editors
@@ -43,7 +39,6 @@ export { LayoutBodyEditor } from "./LayoutBodyEditor";
 export { MenuEditor } from "./MenuEditor";
 export { MenuItemEditor } from "./MenuItemEditor";
 export { NumberFieldEditor } from "./NumberFieldEditor";
-export { SearchFieldEditor } from "./SearchFieldEditor";
 export { ModalEditor } from "./ModalEditor";
 // ⭐ Layout/Slot 관련
 export { SlotEditor } from "./SlotEditor";